+++ backend/common-package/src/main/java/com/ocean/commonPackage/frontParamEntity/chatRoom/friend/UpdateFriendParams.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 // index与deleteFriendParams中的index相同
 public class UpdateFriendParams {
     private String code; // 用户code
-    private Long index; // 好友列表中的索引
+    private Integer index; // 好友列表中的索引
     private Friend friend; // 好友信息
 }
